move smpv from the test into a separate smpv.c
move smpv from the test into a separate smpv.c
