+++ Makefile.am
@@ -64,6 +64,7 @@ noinst_HEADERS = \
     include/dogecoin/serialize.h \
     include/dogecoin/seal.h \
     include/dogecoin/sha2.h \
+    include/dogecoin/smpv.h \
     include/dogecoin/tool.h \
     include/dogecoin/transaction.h \
     include/dogecoin/tx.h \
@@ -112,6 +113,7 @@ libdogecoin_la_SOURCES = \
     src/sign.c \
     src/serialize.c \
     src/sha2.c \
+    src/smpv.c \
     src/cli/tool.c \
     src/transaction.c \
     src/tx.c \
@@ -210,6 +212,7 @@ tests_SOURCES = \
     test/sha1_tests.c \
     test/sha2_tests.c \
     test/signmsg_tests.c \
+    test/smpv_tests.c \
     test/tpm_tests.c \
     test/transaction_tests.c \
     test/tx_tests.c \
