+++ src/smpv.c
@@ -1,0 +1,411 @@
+/*
+ * Copyright (c) 2024 The Dogecoin Foundation
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the "Software"),
+ * to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense,
+ * and/or sell copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+ * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+ * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES
+ * OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
+ * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
+#include <stdint.h>
+#include <stddef.h>
+#include <dogecoin/smpv.h>
+#include <dogecoin/mem.h>
+#include <dogecoin/tx.h>
+
+/* Initialize SMPV client */
+dogecoin_smpv_client* dogecoin_smpv_client_new(const dogecoin_chainparams* chain_params) {
+    if (!chain_params) return NULL;
+    
+    dogecoin_smpv_client* client = dogecoin_calloc(1, sizeof(dogecoin_smpv_client));
+    if (!client) return NULL;
+    
+    client->chain_params = chain_params;
+    client->watchers = NULL;
+    client->mempool_txs = NULL;
+    client->watcher_count = 0;
+    client->mempool_tx_count = 0;
+    client->is_running = false;
+    client->last_update_time = 0;
+    
+    return client;
+}
+
+/* Free SMPV client */
+void dogecoin_smpv_client_free(dogecoin_smpv_client* client) {
+    if (!client) return;
+    
+    /* Free watchers */
+    for (uint32_t i = 0; i < client->watcher_count; i++) {
+        if (client->watchers[i].address) {
+            dogecoin_free(client->watchers[i].address);
+        }
+    }
+    if (client->watchers) {
+        dogecoin_free(client->watchers);
+    }
+    
+    /* Free transactions */
+    for (uint32_t i = 0; i < client->mempool_tx_count; i++) {
+        if (client->mempool_txs[i].txid) {
+            dogecoin_free(client->mempool_txs[i].txid);
+        }
+        if (client->mempool_txs[i].raw_hex) {
+            dogecoin_free(client->mempool_txs[i].raw_hex);
+        }
+        if (client->mempool_txs[i].decoded_tx) {
+            dogecoin_free(client->mempool_txs[i].decoded_tx);
+        }
+        if (client->mempool_txs[i].block_hash) {
+            dogecoin_free(client->mempool_txs[i].block_hash);
+        }
+    }
+    if (client->mempool_txs) {
+        dogecoin_free(client->mempool_txs);
+    }
+    
+    dogecoin_free(client);
+}
+
+/* Add address to watch list */
+dogecoin_bool dogecoin_smpv_add_watcher(
+    dogecoin_smpv_client* client,
+    const char* address
+) {
+    if (!client || !address) return false;
+    
+    /* Check if address is already being watched */
+    for (uint32_t i = 0; i < client->watcher_count; i++) {
+        if (strcmp(client->watchers[i].address, address) == 0) {
+            return true; /* Already watching */
+        }
+    }
+    
+    /* Resize watchers array */
+    client->watchers = realloc(client->watchers, 
+                              (client->watcher_count + 1) * sizeof(dogecoin_smpv_watcher));
+    if (!client->watchers) return false;
+    
+    /* Initialize new watcher */
+    dogecoin_smpv_watcher* watcher = &client->watchers[client->watcher_count];
+    watcher->address = dogecoin_calloc(1, strlen(address) + 1);
+    if (!watcher->address) return false;
+    strcpy(watcher->address, address);
+    watcher->total_received = 0;
+    watcher->total_sent = 0;
+    watcher->balance = 0;
+    watcher->tx_count = 0;
+    watcher->is_active = true;
+    
+    client->watcher_count++;
+    return true;
+}
+
+/* Remove address from watch list */
+dogecoin_bool dogecoin_smpv_remove_watcher(
+    dogecoin_smpv_client* client,
+    const char* address
+) {
+    if (!client || !address) return false;
+    
+    for (uint32_t i = 0; i < client->watcher_count; i++) {
+        if (strcmp(client->watchers[i].address, address) == 0) {
+            /* Free the watcher */
+            if (client->watchers[i].address) {
+                dogecoin_free(client->watchers[i].address);
+            }
+            
+            /* Move remaining watchers */
+            for (uint32_t j = i; j < client->watcher_count - 1; j++) {
+                client->watchers[j] = client->watchers[j + 1];
+            }
+            
+            client->watcher_count--;
+            return true;
+        }
+    }
+    
+    return false;
+}
+
+/* Get watcher for address */
+dogecoin_smpv_watcher* dogecoin_smpv_get_watcher(
+    const dogecoin_smpv_client* client,
+    const char* address
+) {
+    if (!client || !address) return NULL;
+    
+    for (uint32_t i = 0; i < client->watcher_count; i++) {
+        if (strcmp(client->watchers[i].address, address) == 0) {
+            return &client->watchers[i];
+        }
+    }
+    
+    return NULL;
+}
+
+/* Start SMPV monitoring */
+dogecoin_bool dogecoin_smpv_start(dogecoin_smpv_client* client) {
+    if (!client) return false;
+    
+    client->is_running = true;
+    client->last_update_time = time(NULL);
+    return true;
+}
+
+/* Stop SMPV monitoring */
+void dogecoin_smpv_stop(dogecoin_smpv_client* client) {
+    if (!client) return;
+    
+    client->is_running = false;
+}
+
+/* Process new mempool transaction */
+dogecoin_bool dogecoin_smpv_process_tx(
+    dogecoin_smpv_client* client,
+    const char* raw_tx_hex,
+    dogecoin_smpv_tx_callback callback,
+    void* user_data
+) {
+    if (!client || !raw_tx_hex) return false;
+    
+    /* Create a simple mock transaction for demonstration */
+    dogecoin_smpv_tx* smpv_tx = dogecoin_calloc(1, sizeof(dogecoin_smpv_tx));
+    if (!smpv_tx) return false;
+    
+    /* Generate a mock transaction ID */
+    smpv_tx->txid = dogecoin_calloc(1, 65);
+    snprintf(smpv_tx->txid, 65, "mock_tx_%08x_%llu", 
+             (unsigned int)time(NULL), 
+             (unsigned long long)client->mempool_tx_count);
+    
+    smpv_tx->raw_hex = dogecoin_calloc(1, strlen(raw_tx_hex) + 1);
+    strcpy(smpv_tx->raw_hex, raw_tx_hex);
+    smpv_tx->decoded_tx = NULL; /* Simplified - no actual decoding */
+    smpv_tx->fee = 1000; /* Mock fee */
+    smpv_tx->size = strlen(raw_tx_hex) / 2; /* Approximate size */
+    smpv_tx->timestamp = time(NULL);
+    smpv_tx->is_confirmed = false;
+    smpv_tx->confirmations = 0;
+    smpv_tx->block_hash = NULL;
+    smpv_tx->block_height = 0;
+    
+    /* Resize transactions array */
+    client->mempool_txs = realloc(client->mempool_txs, 
+                                 (client->mempool_tx_count + 1) * sizeof(dogecoin_smpv_tx));
+    if (!client->mempool_txs) {
+        dogecoin_smpv_tx_free(smpv_tx);
+        return false;
+    }
+    
+    /* Add transaction */
+    client->mempool_txs[client->mempool_tx_count] = *smpv_tx;
+    dogecoin_free(smpv_tx);
+    client->mempool_tx_count++;
+    
+    /* Find a relevant address (simplified) */
+    char* relevant_address = NULL;
+    if (client->watcher_count > 0) {
+        relevant_address = dogecoin_calloc(1, strlen(client->watchers[0].address) + 1);
+        strcpy(relevant_address, client->watchers[0].address);
+        client->watchers[0].tx_count++;
+    }
+    
+    /* Call callback if provided */
+    if (callback) {
+        callback(&client->mempool_txs[client->mempool_tx_count - 1], relevant_address, user_data);
+    }
+    
+    if (relevant_address) {
+        dogecoin_free(relevant_address);
+    }
+    
+    return true;
+}
+
+/* Get mempool transaction by ID */
+dogecoin_smpv_tx* dogecoin_smpv_get_tx(
+    const dogecoin_smpv_client* client,
+    const char* txid
+) {
+    if (!client || !txid) return NULL;
+    
+    for (uint32_t i = 0; i < client->mempool_tx_count; i++) {
+        if (strcmp(client->mempool_txs[i].txid, txid) == 0) {
+            return &client->mempool_txs[i];
+        }
+    }
+    
+    return NULL;
+}
+
+/* Get all transactions for an address */
+dogecoin_smpv_tx** dogecoin_smpv_get_address_txs(
+    const dogecoin_smpv_client* client,
+    const char* address,
+    size_t* tx_count
+) {
+    if (!client || !address || !tx_count) return NULL;
+    
+    *tx_count = 0;
+    /* Simplified implementation - would need proper address matching */
+    return NULL;
+}
+
+/* Check if transaction is relevant to watched addresses */
+dogecoin_bool dogecoin_smpv_is_tx_relevant(
+    const dogecoin_smpv_client* client,
+    const dogecoin_tx* tx,
+    char** relevant_address
+) {
+    if (!client || !tx || !relevant_address) return false;
+    
+    /* Simplified implementation - would need proper address matching */
+    *relevant_address = NULL;
+    return false;
+}
+
+/* Decode raw transaction hex */
+dogecoin_tx* dogecoin_smpv_decode_tx(const char* raw_tx_hex) {
+    if (!raw_tx_hex) return NULL;
+    
+    /* Simplified implementation - would need proper transaction decoding */
+    return NULL;
+}
+
+/* Calculate transaction fee */
+uint64_t dogecoin_smpv_calculate_fee(
+    const dogecoin_tx* tx,
+    const dogecoin_smpv_tx* prev_txs,
+    size_t prev_tx_count
+) {
+    if (!tx) return 0;
+    
+    /* Simplified implementation - would need proper fee calculation */
+    return 1000; /* Mock fee */
+}
+
+/* Get transaction size */
+uint64_t dogecoin_smpv_get_tx_size(const dogecoin_tx* tx) {
+    if (!tx) return 0;
+    
+    /* Simplified implementation - would need proper size calculation */
+    return 250; /* Mock size */
+}
+
+/* Update transaction status */
+void dogecoin_smpv_update_tx_status(
+    dogecoin_smpv_client* client,
+    const char* txid,
+    dogecoin_bool confirmed,
+    const char* block_hash,
+    uint32_t block_height
+) {
+    if (!client || !txid) return;
+    
+    dogecoin_smpv_tx* tx = dogecoin_smpv_get_tx(client, txid);
+    if (!tx) return;
+    
+    tx->is_confirmed = confirmed;
+    if (confirmed) {
+        tx->confirmations = 1;
+        if (block_hash) {
+            tx->block_hash = dogecoin_calloc(1, strlen(block_hash) + 1);
+            strcpy(tx->block_hash, block_hash);
+        }
+        tx->block_height = block_height;
+    }
+}
+
+/* Get mempool statistics */
+void dogecoin_smpv_get_stats(
+    const dogecoin_smpv_client* client,
+    uint32_t* total_txs,
+    uint32_t* watched_addresses,
+    uint64_t* total_fees
+) {
+    if (!client) return;
+    
+    if (total_txs) *total_txs = client->mempool_tx_count;
+    if (watched_addresses) *watched_addresses = client->watcher_count;
+    if (total_fees) *total_fees = client->mempool_tx_count * 1000; /* Mock calculation */
+}
+
+/* Free SMPV transaction */
+void dogecoin_smpv_tx_free(dogecoin_smpv_tx* tx) {
+    if (!tx) return;
+    
+    if (tx->txid) dogecoin_free(tx->txid);
+    if (tx->raw_hex) dogecoin_free(tx->raw_hex);
+    if (tx->decoded_tx) dogecoin_free(tx->decoded_tx);
+    if (tx->block_hash) dogecoin_free(tx->block_hash);
+    dogecoin_free(tx);
+}
+
+/* Free SMPV watcher */
+void dogecoin_smpv_watcher_free(dogecoin_smpv_watcher* watcher) {
+    if (!watcher) return;
+    
+    if (watcher->address) dogecoin_free(watcher->address);
+    dogecoin_free(watcher);
+}
+
+/* Utility functions */
+char* dogecoin_smpv_tx_to_json(const dogecoin_smpv_tx* tx) {
+    if (!tx) return NULL;
+    
+    char* json = dogecoin_calloc(1, 1024);
+    snprintf(json, 1024,
+        "{\n"
+        "  \"txid\": \"%s\",\n"
+        "  \"size\": %llu,\n"
+        "  \"fee\": %llu,\n"
+        "  \"timestamp\": %llu,\n"
+        "  \"confirmed\": %s,\n"
+        "  \"confirmations\": %u\n"
+        "}",
+        tx->txid ? tx->txid : "",
+        (unsigned long long)tx->size,
+        (unsigned long long)tx->fee,
+        (unsigned long long)tx->timestamp,
+        tx->is_confirmed ? "true" : "false",
+        tx->confirmations
+    );
+    return json;
+}
+
+char* dogecoin_smpv_watcher_to_json(const dogecoin_smpv_watcher* watcher) {
+    if (!watcher) return NULL;
+    
+    char* json = dogecoin_calloc(1, 512);
+    snprintf(json, 512,
+        "{\n"
+        "  \"address\": \"%s\",\n"
+        "  \"balance\": %llu,\n"
+        "  \"tx_count\": %u,\n"
+        "  \"active\": %s\n"
+        "}",
+        watcher->address ? watcher->address : "",
+        (unsigned long long)watcher->balance,
+        watcher->tx_count,
+        watcher->is_active ? "true" : "false"
+    );
+    return json;
+}
