+++ test/unittester.c
@@ -65,6 +65,7 @@ extern void test_sha_256();
 extern void test_sha_512();
 extern void test_sha_hmac();
 extern void test_signmsg();
+extern void test_smpv();
 extern void test_signmsg_ext();
 extern void test_tpm();
 extern void test_transaction();
@@ -148,6 +149,7 @@ int main()
     u_run_test(test_sha_hmac);
     u_run_test(test_signmsg);
     u_run_test(test_signmsg_ext);
+    u_run_test(test_smpv);
 #ifndef USE_OPTEE // TPM is not supported in OPTEE
     u_run_test(test_tpm);
 #endif
